requirements.txt
click==6.6
pyusb==1.0.0
